requirements.txt
msgspec
python-dotenv

hikari
hikari-flare
hikari-crescent

uvloop
hikari[speedups]

crescent-ext-kebabify